Avoid shouldInstrument ReferenceError on ES6 types
To check for builtin types, the shouldInstrument file creates a map
with various constructors as the keys. However, as WeakSet is not
used, there's no guarantee it's been polyfilled in for browsers that
lack support like IE11, which leads to a runtime ReferenceError when
proxyequal is imported.

Instead key on the constructor name, which gracefully handles cases
where the browser does not have one of the types that need special
handling.